requirements.txt
opencv-python
face_recognition
mysql-connector-python
numpy
Pillow
